functions/requirements.txt
firebase_functions~=0.1.0
functions-framework==3.*
firebase-functions
firebase-admin
weasyprint
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
python-multipart